Bugfix: EF_BUILD_RANDOM buffer overrun disables I2C buttons.
The Animation parameter initialization for `Effect::EF_BUILD_RANDOM`
cleared out a memory array using `memset()` that zeroed out far past
the end of the array.

(At least one of) the effects of this is that the I2C communication
with the 8 I2C-based buttons on the button grid would silently fail
for the remainder of the sketch execution, after this animation started.

Reset `memset()` size parameter to use correct type (uint8_t) for the
`Animation::_buildRandomOrder` buffer.

+++ animation.cpp
@@ -216,7 +216,7 @@ void Animation::_setParamsBuildRandom(const Sentence &s, const Effect e, uint32_
   // Randomize the order we light up the sentence words in this animation session:
 
   // Step 1: Reset ordering array.
-  memset(_buildRandomOrder, 0, sizeof(unsigned int) * NUM_SIGNS);
+  memset(_buildRandomOrder, 0, sizeof(uint8_t) * NUM_SIGNS);
 
   // Step 2: Put sign ids into the ordering array sequentially.
   uint32_t signBits = s.getSignBits();
